Crate air_interpreter_interface

Structs

CallArgumentsRepr

CallRequestParams

Contains arguments of a call instruction and all other necessary information required for calling a service.

CallRequestsRepr

CallResultsRepr

CallServiceResult

Represents an executed host function result.

InterpreterOutcome

Describes a result returned at the end of the interpreter execution_step.

RunParameters

Parameters that a host side should pass to an interpreter and that necessary for execution.

SerializedCallArguments

SerializedCallRequests

SerializedCallResults

SerializedTetraplets

SoftLimitsTriggering

This stores soft limits triggering flags.

TetrapletsRepr

Constants

CALL_SERVICE_SUCCESS

INTERPRETER_SUCCESS

Statics

MAX_AIR_SIZE

These are RAM consumption related limits to be enforced by AquaVM. There are two enforcing modes in AquaVM: soft and hard limit. The mode is signalled by AquaVM function Invoker via its run parameters. Soft limit mode sets a set of flags to return to the Invoker. Hard limit mode forces AquaVM to return Uncatchable error if the limits are exceeded. The math behind the limits value is based on:

MAX_CALL_RESULT_SIZE

MAX_PARTICLE_SIZE

Functions

try_as_string

Type Aliases

CallArgumentsDeserializeError

CallArgumentsFormat

CallRequests

CallRequestsDeserializeError

CallRequestsFormat

CallResults

This is a map from a String to a service result for compatibility with JavaScript. Binary format implementations like rmp-serde do not convert keys from strings, unlike serde_json.

CallResultsDeserializeError

CallResultsFormat

CallResultsSerializeError

TetrapletDeserializeError

TetrapletsFormat